requirements.txt
flask>=2.3.0
flask_sqlalchemy>=3.1.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
scikit-learn>=1.2.0
pytest>=7.0.0